Fix some issues with decimal values in ingredients
Also cleaned up some controller code and UI tweaks

+++ lib/precider/catalog/product_ingredient.ex
@@ -5,7 +5,7 @@ defmodule Precider.Catalog.ProductIngredient do
   alias Precider.Catalog.{Product, Ingredient}
 
   schema "product_ingredients" do
-    field :dosage_amount, :integer
+    field :dosage_amount, :decimal
     field :dosage_unit, Ecto.Enum, values: [:mg, :g, :mcg]
     
     belongs_to :product, Product
